refactor: create cutom hook in today
